image/test/mochitest/test_animSVGImage.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=610419
-->
<head>
  <title>Test for Bug 610419</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <script type="application/javascript" src="imgutils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=610419">Mozilla Bug 610419</a>
<p id="display"></p>
<div id="content">
  <div id="referenceDiv" style="height: 100px; width: 100px;
                                display: none; background: lime"></div>
  <img>
</div>
<pre id="test">
<script type="application/javascript;version=1.8">
/** Test for Bug 610419**/

SimpleTest.requestFlakyTimeout("Pre-existing timeouts when converting from mochitest-chrome");
SimpleTest.waitForExplicitFinish();

const FAILURE_TIMEOUT = 120000; // Fail early after 120 seconds (2 minutes)

const gImg = document.getElementsByTagName("img")[0];

var gMyDecoderObserver; // value will be set in main()
var gReferenceSnapshot; // value will be set in takeReferenceSnapshot()
var gPollCounter = 0;
var gIsTestFinished = false;
var gSVGImages = [
  "lime-anim-100x100.svg",    // SMIL animation
  "lime-css-anim-100x100.svg" // CSS animation
]
var gSVGCurrentImage = 0;

function takeReferenceSnapshot() {
  // Take a snapshot of the initial (essentially blank) page
  let blankSnapshot = snapshotWindow(window, false);

  // Show reference div, & take a snapshot
  let referenceDiv = document.getElementById("referenceDiv");
  referenceDiv.style.display = "block";
  gReferenceSnapshot = snapshotWindow(window, false);
  ok(compareSnapshots(blankSnapshot, gReferenceSnapshot, false)[0],
     "reference snapshot shouldn't match blank page snapshot");

  // Re-hide reference div, and take another snapshot to be sure it's gone
  referenceDiv.style.display = "none";
  let blankSnapshot2 = snapshotWindow(window, false);
  ok(compareSnapshots(blankSnapshot, blankSnapshot2, true)[0],
     "reference div should disappear when it becomes display:none");
}

function loadNextImageAndPoll()
{
  setTimeout(myPoll, 1);
  // kick off image-loading! myPoll handles the rest.
  gImg.setAttribute("src", gSVGImages[gSVGCurrentImage]);
}

function myPoll() {
  gPollCounter++;
  ok(true, "myPoll called");
  let currentSnapshot = snapshotWindow(window, false);
  if (compareSnapshots(currentSnapshot, gReferenceSnapshot, true)[0]) {
    // SUCCESS!
    ok(true, "Animated image looks correct, " +
             "at call #" + gPollCounter + " to myPoll");

    if (++gSVGCurrentImage > gSVGImages.length) {
      cleanUpAndFinish();
    } else {
      loadNextImageAndPoll();
    }
  }
  else
    setTimeout(myPoll, 1);
}

function failTest() {
  ok(false, "timing out after " + FAILURE_TIMEOUT + "ms.  " +
            "Animated image still doesn't look correct, " +
            "after call #" + gPollCounter + " to myPoll");
  cleanUpAndFinish();
}

function cleanUpAndFinish() {
  // On the off chance that failTest and myPoll are triggered
  // back-to-back, use a flag to prevent multiple calls to SimpleTest.finish.
  if (gIsTestFinished) {
    return;
  }
  SimpleTest.finish();
  gIsTestFinished = true;
}

function main() {
  takeReferenceSnapshot();

  // We want to test the cold loading behavior, so clear cache in case an
  // earlier test got our image in there already.
  clearAllImageCaches();

  loadNextImageAndPoll();

  // In case something goes wrong, fail earlier than mochitest timeout,
  // and with more information.
  setTimeout(failTest, FAILURE_TIMEOUT);
}

window.onload = main;

</script>
</pre>
</body>
</html>
